Prepares for deployment to AWS
