+++ database/pg-schema2.sql
@@ -1,0 +1,18 @@
+DROP TABLE IF EXISTS games;
+
+CREATE TABLE games (
+    proxyid integer PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
+    game jsonb);
+
+COPY games(game) FROM '/home/jay/Documents/Projects/sidebar-body-related/database/datajson.csv' with (format csv, quote '|', delimiter ';');
+
+-- COPY games FROM '/home/jay/Documents/Projects/sidebar-body-related/database/data.json';
+
+/*  
+
+    When in databases directory in terminal:
+        > psql steamy
+
+    From postgres cli:
+        steamy=# \i 'pg-schema2.sql';
+*/
