+++ package.json
@@ -8,7 +8,7 @@
     "start": "nodemon ./server/server.js -i client",
     "babel": "babel client --out-dir public --presets=env,react --source-maps inline --watch",
     "build": "webpack --watch",
-    "seed": "node ./database/seedscript.js",
+    "seed": "node ./database/seedpostgres.js",
     "jest": "jest --detectOpenHandles --coverage --watch",
     "lint": "eslint ./",
     "lint-client": "eslint ./client --ext .js,.jsx",
@@ -16,21 +16,23 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/Team-Gandalf/sidebar-body-related.git"
+    "url": "git+https://github.com/SDC-HRR/Jay-Service"
   },
-  "author": "Haider Syed",
+  "author": "Jay Barr",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/Team-Gandalf/sidebar-body-related/issues"
+    "url": "https://github.com/SDC-HRR/Jay-Service/issues"
   },
-  "homepage": "https://github.com/Team-Gandalf/sidebar-body-related#readme",
+  "homepage": "hhttps://github.com/SDC-HRR/Jay-Service#readme",
   "dependencies": {
     "axios": "^0.19.2",
     "body-parser": "^1.19.0",
+    "cassandra-driver": "^4.5.2",
     "express": "^4.17.1",
     "faker": "^4.1.0",
     "mongoose": "^5.9.15",
     "path": "^0.12.7",
+    "pg-promise": "^10.5.6",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-slick": "^0.25.2",
